refactor: 높이 검증 기능을 높이 객체로 이동
